Server accepts concurrent clients, Client can send any text input over to server, server can send response back
I opened a new branch for this as this is a bit of refactoring for both client and server, will need to be merged with the other branch as it felt a bit messy to work in there.

Client can now send any textual input to the Server and will then receive a response back. The server currently responds to "exit, buyer, seller" and will send a corresponding response back or end the connection.

Main.c is no longer necessary, Server.c includes its own main function to launch itself now.